+++ changelog.md
@@ -7,3 +7,4 @@
 ## Next
 
 * `.readEntries` + `.walkEntries` avoid creating long promise chains (closes #2)
+* `.walkEntries` awaits promise from callback (closes #1)